# pkg/textparse: refactor and add metadata handling