… pkg/textparse/lex.l
… pkg/textparse/lex.l.go