Extends the parser to allow retrieving metadata.
The lexer now yields proper tokens that are fed into a hand-written
parser on top.

@jkohen @brian-brazil 

```
benchmark                                                    old ns/op     new ns/op     delta
BenchmarkParse/no-decode-metric/testdata.txt-8               266           270           +1.50%
BenchmarkParse/decode-metric/testdata.txt-8                  432           427           -1.16%
BenchmarkParse/decode-metric-reuse/testdata.txt-8            355           356           +0.28%
BenchmarkParse/no-decode-metric/testdata.nometa.txt-8        209           219           +4.78%
BenchmarkParse/decode-metric/testdata.nometa.txt-8           372           372           +0.00%
BenchmarkParse/decode-metric-reuse/testdata.nometa.txt-8     300           304           +1.33%
BenchmarkGzip/testdata.txt-8                                 441           447           +1.36%
BenchmarkGzip/testdata.nometa.txt-8                          287           289           +0.70%

benchmark                                                    old allocs     new allocs     delta
BenchmarkParse/no-decode-metric/testdata.txt-8               0              0              +0.00%
BenchmarkParse/decode-metric/testdata.txt-8                  1              1              +0.00%
BenchmarkParse/decode-metric-reuse/testdata.txt-8            1              1              +0.00%
BenchmarkParse/no-decode-metric/testdata.nometa.txt-8        0              0              +0.00%
BenchmarkParse/decode-metric/testdata.nometa.txt-8           1              1              +0.00%
BenchmarkParse/decode-metric-reuse/testdata.nometa.txt-8     1              1              +0.00%
BenchmarkGzip/testdata.txt-8                                 0              0              +0.00%
BenchmarkGzip/testdata.nometa.txt-8                          0              0              +0.00%

benchmark                                                    old bytes     new bytes     delta
BenchmarkParse/no-decode-metric/testdata.txt-8               0             0             +0.00%
BenchmarkParse/decode-metric/testdata.txt-8                  128           127           -0.78%
BenchmarkParse/decode-metric-reuse/testdata.txt-8            48            47            -2.08%
BenchmarkParse/no-decode-metric/testdata.nometa.txt-8        0             0             +0.00%
BenchmarkParse/decode-metric/testdata.nometa.txt-8           128           127           -0.78%
BenchmarkParse/decode-metric-reuse/testdata.nometa.txt-8     48            47            -2.08%
BenchmarkGzip/testdata.txt-8                                 416           416           +0.00%
BenchmarkGzip/testdata.nometa.txt-8                          256           256           +0.00%
```

Ultimately performance remains the same as before.