@fabxc 2018-05-25
Addressed all those and added tests. Had to resort to another little work around to not make comment handling overly complex. But test coverage is pretty good so that should be fine.