@fabxc 2018-06-01
What's the deal with this Travis flake? I've seen it many times the last ~2 weeks:

```
>> running staticcheck
/home/travis/gopath/bin/staticcheck -ignore "github.com/prometheus/prometheus/discovery/kubernetes/kubernetes.go:SA1019 github.com/prometheus/prometheus/discovery/kubernetes/node.go:SA1019 github.com/prometheus/prometheus/documentation/examples/remote_storage/remote_storage_adapter/main.go:SA1019 github.com/prometheus/prometheus/pkg/textparse/lex.l.go:SA4006 github.com/prometheus/prometheus/pkg/pool/pool.go:SA6002 github.com/prometheus/prometheus/promql/engine.go:SA6002" ./...
Killed
make: *** [staticcheck] Error 137
```