@fabxc 2018-06-01
It seems to always happen in the staticcheck part. Since we are running everything sequentially, the race detector in the previous steps shouldn't matter. Maybe staticcheck just gets bloated with all our dependencies.